Remove race condition in ldapauth.c
1. Remove calling `is_valid_socket()` in `ldap_worker_thread` thread.
This could not impact the final result of authentication, because client->state
could only be changed after timeout, so `disconnect_client()` will be called
firstly, then if ldap authentication succeed, it will call
`sbuf_continue(&client->sbuf)`, but nothing happens because the client is disconnected.
2. Add mutex to protect accessing `client->status` in main thread and worker thread.
The main thread will check the value of `client->status` in `ldap_poll()`, and the
`ldap_worker_thread` thread will set the value of `client->status`. So we add a mutex
in each of a request to protect it from race condition.